# src/api/rest-api/openapi.json
{
  "components": {
    "schemas": {
...
      "BranchModel": {
        "properties": {
          "createdAt": {
            "description": "Timestamp when the branch was created",
            "example": "2024-01-15T10:30:00.000Z",
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "description": "Unique branch identifier",
            "example": "V1StGXR8_Z5jdHi6B-myT",
            "type": "string"
          },
          "isRoot": {
            "description": "Whether this is the root (default) branch of the project. Each project has exactly one root branch.",
            "type": "boolean"
          },
          "name": {
            "description": "Branch name",
            "example": "master",
            "type": "string"
          },
          "projectId": {
            "description": "Project this branch belongs to",
            "example": "X2TuHYS9_A6keJj7C-nzU",
            "type": "string"
          }
        },
        "required": [
          "id",
          "projectId",
          "createdAt",
          "isRoot",
          "name"
        ],
        "type": "object"
      },
...
      "CreateEndpointDto": {
        "properties": {
          "type": {
            "enum": [
              "GRAPHQL",
              "REST_API"
            ],
            "type": "string"
          }
        },
        "required": [
          "type"
        ],
        "type": "object"
      },
...
      "EndpointModel": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "type": {
            "enum": [
              "GRAPHQL",
              "REST_API"
            ],
            "type": "string"
          }
        },
        "required": [
          "id",
          "createdAt",
          "type"
        ],
        "type": "object"
      },
      "ErrorModel": {
        "properties": {
          "error": {
            "description": "Error type",
            "example": "Bad Request",
            "type": "string"
          },
          "message": {
            "description": "Detailed error message",
            "example": "Validation failed",
            "type": "string"
          },
          "statusCode": {
            "description": "HTTP status code",
            "example": 400,
            "type": "number"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ],
        "type": "object"
      },
      "GetEndpointResultDto": {
        "properties": {
          "branch": {
            "$ref": "#/components/schemas/BranchModel"
          },
          "endpoint": {
            "$ref": "#/components/schemas/EndpointModel"
          },
          "project": {
            "$ref": "#/components/schemas/ProjectModel"
          },
          "revision": {
            "$ref": "#/components/schemas/RevisionModel"
          }
        },
        "required": [
          "endpoint",
          "revision",
          "branch",
          "project"
        ],
        "type": "object"
      },
...
      "ProjectModel": {
        "properties": {
          "createdAt": {
            "description": "Timestamp when the project was created",
            "example": "2024-01-15T10:30:00.000Z",
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "description": "Unique project identifier",
            "example": "V1StGXR8_Z5jdHi6B-myT",
            "type": "string"
          },
          "name": {
            "description": "URL-friendly project name (lowercase letters, numbers, hyphens)",
            "example": "my-blog",
            "type": "string"
          },
          "organizationId": {
            "description": "Organization that owns this project",
            "example": "acme-corp",
            "type": "string"
          }
        },
        "required": [
          "id",
          "organizationId",
          "createdAt",
          "name"
        ],
        "type": "object"
      },
...
      "RevisionModel": {
        "properties": {
          "createdAt": {
            "description": "Timestamp when the revision was created",
            "example": "2024-01-15T10:30:00.000Z",
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "description": "Unique revision identifier",
            "example": "V1StGXR8_Z5jdHi6B-myT",
            "type": "string"
          },
          "isDraft": {
            "description": "Whether this is the draft revision. Draft revisions can be modified; use for all write operations.",
            "type": "boolean"
          },
          "isHead": {
            "description": "Whether this is the head (latest committed) revision. Head is immutable and represents the published state.",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "createdAt",
          "isDraft",
          "isHead"
        ],
        "type": "object"
      },
...
      "SuccessModelDto": {
        "properties": {
          "success": {
            "example": true,
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "type": "object"
      },
...
      "UpdateProjectDto": {
        "properties": {
          "isPublic": {
            "type": "boolean"
          }
        },
        "required": [
          "isPublic"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "access-token": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "",
    "title": "Revisium API",
    "version": "2.6.0-alpha.3"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/endpoints/{endpointId}": {
      "delete": {
        "operationId": "deleteEndpoint",
        "parameters": [
          {
            "description": "Endpoint identifier",
            "in": "path",
            "name": "endpointId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Invalid request data or validation failed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Authentication required. Provide a valid JWT token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Insufficient permissions for this operation"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Endpoint not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete an endpoint",
        "tags": [
          "Endpoint"
        ]
      }
    },
    "/api/endpoints/{endpointId}/relatives": {
      "get": {
        "operationId": "endpointRelatives",
        "parameters": [
          {
            "description": "Endpoint identifier",
            "in": "path",
            "name": "endpointId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetEndpointResultDto"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Invalid request data or validation failed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Authentication required. Provide a valid JWT token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Insufficient permissions for this operation"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Endpoint not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get endpoint with all related entities",
        "tags": [
          "Endpoint"
        ]
      }
    },
...
    "/api/organization/{organizationId}/projects/{projectName}": {
      "delete": {
        "operationId": "deleteProject",
        "parameters": [
          {
            "description": "Organization identifier (typically the owner username)",
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "URL-friendly project name (lowercase, numbers, hyphens allowed)",
            "in": "path",
            "name": "projectName",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessModelDto"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Invalid request data or validation failed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Authentication required. Provide a valid JWT token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Insufficient permissions for this operation"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Project not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete a project",
        "tags": [
          "Project"
        ]
      },
      "get": {
        "operationId": "project",
        "parameters": [
          {
            "description": "Organization identifier (typically the owner username)",
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "URL-friendly project name (lowercase, numbers, hyphens allowed)",
            "in": "path",
            "name": "projectName",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectModel"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Invalid request data or validation failed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Authentication required. Provide a valid JWT token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Insufficient permissions for this operation"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Project not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get project by name",
        "tags": [
          "Project"
        ]
      },
      "put": {
        "operationId": "updateProject",
        "parameters": [
          {
            "description": "Organization identifier (typically the owner username)",
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "URL-friendly project name (lowercase, numbers, hyphens allowed)",
            "in": "path",
            "name": "projectName",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessModelDto"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Invalid request data or validation failed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Authentication required. Provide a valid JWT token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Insufficient permissions for this operation"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Project not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update project settings",
        "tags": [
          "Project"
        ]
      }
    },
...
    "/api/revision/{revisionId}/endpoints": {
      "get": {
        "operationId": "endpoints",
        "parameters": [
          {
            "description": "Revision identifier (any revision: draft, head, or historical)",
            "in": "path",
            "name": "revisionId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EndpointModel"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Invalid request data or validation failed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Authentication required. Provide a valid JWT token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Insufficient permissions for this operation"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Revision not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List API endpoints for this revision",
        "tags": [
          "Revision"
        ]
      },
      "post": {
        "operationId": "createEndpoint",
        "parameters": [
          {
            "description": "Revision identifier (any revision: draft, head, or historical)",
            "in": "path",
            "name": "revisionId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEndpointDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EndpointModel"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Invalid request data or validation failed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Authentication required. Provide a valid JWT token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Insufficient permissions for this operation"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Revision not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create an API endpoint for this revision",
        "tags": [
          "Revision"
        ]
      }
    },
...
  },
  "servers": [],
...
}